Adds seccomp whitelisting
With the provided whitelist team-suse is limited to only the defined syscalls.
Everything else is not permitted after the rules are added. This has the
purpose to limit the attack surface through this program.

whitelist.py is a script to quickly format the syscalls collected with strace.

Additional dependencies where vendored with dep.