Adds seccomp whitelisting (#1)
* Adds seccomp whitelisting

With the provided whitelist team-suse is limited to only the defined syscalls.
Everything else is not permitted after the rules are added. This has the
purpose to limit the attack surface through this program.

whitelist.py is a script to quickly format the syscalls collected with strace.

Additional dependencies where vendored with dep.


* Using build constraints for seccomp

Since seccomp can't even be compiled on other platforms, gating has to be
done via build constriants. The seccomp part can now be found in a new
file with a suffix for linux. This way only linux systems are compiled
with the seccomp extension.